Add config option for linking zstd static
Add `--with-static-zstd` option to link zstd library staticly. This option is
mutually exclusive with the original option `--with-zstd`. The original behavior
of `--with-zstd` is not affected.

We need to use `--without-zstd --with-static-zstd` to turn on the option since
`--with-zstd` is enabled by default.

If we link zstd staticly, we can customized allocator for zstd functions. Those
functions are only exposed in the experimental feature. I guess those function
signatures are not likely to change.

Co-authored-by: Dennis Kovalenko <kds@arenadata.io>
Co-authored-by: Chen Mulong <chenmulong@gmail.com>